🦄 refactor(dailynote): 把daily note相关的功能函数移到 dailynote.ts 内
